+++ uspace/lib/c/include/device/hw_res.h
@@ -115,7 +115,7 @@ extern int hw_res_get_resource_list(async_sess_t *, hw_resource_list_t *);
 extern bool hw_res_enable_interrupt(async_sess_t *);
 
 extern int hw_res_dma_channel_setup(async_sess_t *, unsigned int, uint32_t,
-    uint16_t, uint8_t);
+    uint32_t, uint8_t);
 extern int hw_res_dma_channel_remain(async_sess_t *, unsigned);
 
 #endif
